+++ CHANGELOG.md
@@ -1,2 +1,2 @@
-* Fixed Squeeze Leaves loot table
-* Added "Convert Void Holes" config setting
+* Fixed Squeeze Saplings not growing naturally
+* Removed Project Table debug code